Refactored Command, Response, Fixed abridged graphs for response paths
