+++ clients/java-client-lib-alternate/client-core/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>net.zscript</groupId>
-        <artifactId>zscript-clients</artifactId>
+        <artifactId>java-client-lib-alternate</artifactId>
         <version>0.0.1-SNAPSHOT</version>
         <relativePath>../../pom.xml</relativePath>
     </parent>
@@ -13,32 +13,5 @@
     <artifactId>client-core</artifactId>
     <name>Java Client Core</name>
 
-    <dependencies>
-        <dependency>
-            <groupId>net.zscript</groupId>
-            <artifactId>java-receiver-token</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>net.zscript</groupId>
-            <artifactId>zscript-model-schema</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>net.zscript</groupId>
-            <artifactId>zscript-model-standard-module-definitions</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.junit.jupiter</groupId>
-            <artifactId>junit-jupiter-params</artifactId>
-            <version>5.10.0-M1</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.assertj</groupId>
-            <artifactId>assertj-core</artifactId>
-        </dependency>
-    </dependencies>
 
 </project>
