+++ clients/java-client-lib-alternate/pom.xml
@@ -17,6 +17,11 @@
     <dependencies>
         <dependency>
             <groupId>net.zscript</groupId>
+            <artifactId>zscript-model-schema</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>net.zscript</groupId>
             <artifactId>zscript-model-standard-module-definitions</artifactId>
             <version>${project.version}</version>
         </dependency>
@@ -35,26 +40,17 @@
             <artifactId>slf4j-api</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.junit.jupiter</groupId>
-            <artifactId>junit-jupiter</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.junit.vintage</groupId>
-            <artifactId>junit-vintage-engine</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.assertj</groupId>
-            <artifactId>assertj-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.mockito</groupId>
-            <artifactId>mockito-junit-jupiter</artifactId>
-        </dependency>
-        <dependency>
             <groupId>nl.jqno.equalsverifier</groupId>
             <artifactId>equalsverifier</artifactId>
             <version>3.15.1</version>
             <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>net.zscript</groupId>
+            <artifactId>java-testing-deps</artifactId>
+            <version>${project.version}</version>
+            <scope>test</scope>
+            <type>pom</type>
         </dependency>
     </dependencies>
 </project>
